Carpool server and updated Android app
Added server, sql database and user accounts. Android app can log in
and start a new session with the server and periodically update the
user’s location in the database. The NetBeans project for the server is
in the CarpoolServer folder. To run, start MySQL on your machine first,
then run the CarpoolServer project, then finally run the Android
project. In the running Android program you can change the IP:Port of
the server to contact with the Settings button. If running on the phone
this will be the machine’s local network IP address (connected via wifi
or whatever) and not “localhost”. If running through the emulator it
will be the 10.0.2.2:8080 address. The server will create its own
database & tables the first time you create a new user account from the
Android app. In the CarpoolServer project the database config is read
from config-local.xml (or config-online.xml), set at the top of the
CarpoolerEJB.java source - if you need to change it, eg if anybody
wants to try deploying it on to the AUT server.